refactor: Moved starting code into lib with only a thin wrapper in the `main` for building the binary
